casimir Options

Internally used, package-specific options. All options will prioritise R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.

Checking Option Values

Option values specific to casimir can be accessed by passing the package name to env.

options::opts(env = "casimir")

options::opt(x, default, env = "casimir")

Options

ignore_inconsistencies

Warnings about data inconsistencies will be silenced.
default:

FALSE

option:

casimir.ignore_inconsistencies

envvar:

R_CASIMIR_IGNORE_INCONSISTENCIES (evaluated if possible, raw string otherwise)

progress

Display progress bars for iterated computations (like bootstrap CI or pr curves).
default:

FALSE

option:

casimir.progress

envvar:

R_CASIMIR_PROGRESS (evaluated if possible, raw string otherwise)

verbose

Verbose reporting of computation steps for debugging.
default:

FALSE

option:

casimir.verbose

envvar:

R_CASIMIR_VERBOSE (evaluated if possible, raw string otherwise)

check_group_names

Perform replacement of dots in grouping columns. Disable for faster computation if you can make sure that all columns used for grouping ("doc_id", "label_id", "doc_groups", "label_groups") do not contain dots.
default:

TRUE

option:

casimir.check_group_names

envvar:

R_CASIMIR_CHECK_GROUP_NAMES (evaluated if possible, raw string otherwise)

drop_empty_groups

Should empty levels of factor variables be dropped in grouped set retrieval computation?
default:

TRUE

option:

casimir.drop_empty_groups

envvar:

R_CASIMIR_DROP_EMPTY_GROUPS (evaluated if possible, raw string otherwise)

replace_zero_division_with

In macro averaged results (doc-avg, subj-avg), it may occur that some instances have no predictions or no gold standard. In these cases, calculating precision and recall may lead to division by zero. CASIMiR standardly removes these missing values from macro averages, leading to a smaller support (count of instances that were averaged). Other implementations of macro averaged precision and recall default to 0 in these cases. This option allows to control the default. Set any value between 0 and 1.
default:

NULL

option:

casimir.replace_zero_division_with

envvar:

R_CASIMIR_REPLACE_ZERO_DIVISION_WITH (evaluated if possible, raw string otherwise)

See also